Example working!
Now I need to refactor so that I can test with my test cases.

Note though that, as I had feared, there is an issue with rounding such
that we are off by 1 on some of these percentages. I have done what I
can to try and mitigate by doing so on the larger percentage models so
that the overall effect is less, but it is an annoying thing that I
would like to remove if possible.

To that end, I have already seen if ceil and floor make a difference
when creating the alpha, but it does not appear so.